ADMINUI-791 Fixes provisioning POST request results in a http 302 redirect to http, not https
This is caused by Sinatra not konwing what scheme to use when assembling
the URL. When nginx proxies the request to the backend, it's done via
the un SSL'd variant (as thats how the backend processes these
requests). In order to fix this, Rack::Request#ssl? needs to return true
in order for sinatra to correctly assemle URL redirects. One of the
conditions checked by Rack::Request#ssl? is the X-Forwarded-Proto
Header, if this header is set, the protocol specified in the header will
be used as the scheme portiion of the resulting URL for redirects.